+++ src/structured_data/_patterns/bind.py
@@ -1,4 +1,5 @@
 from .._not_in import not_in
+from .._pep_570_when import pep_570_when
 from .basic_patterns import Pattern
 from .compound_match import CompoundMatch
 
@@ -17,8 +18,8 @@ class Bind(CompoundMatch, tuple):
 
     __slots__ = ()
 
-    def __new__(*args, **kwargs):
-        cls, structure = args
+    @pep_570_when
+    def __new__(cls, structure, kwargs):
         if not kwargs:
             return structure
         not_in(kwargs, "_")